Add autocompletion in Cpp mode
a drop down menu with the suggestions simply appears on the screen.
This is achieved matching the autocompleter of ipython and TTabCom.
There is some gymnastic to extract the list of suggestions as container
of names and presently it's done in a jitted function. One could
imagine to add a method to TTabCom in Cpp directly.